intellij init

+++ nadal_1/harjutus_9.java
@@ -4,10 +4,10 @@ public class harjutus_9 {
     public static void main(String[] args) {
         try (Scanner reader = new Scanner(System.in)) {
             System.out.print("Type a number: ");
-            int num1 = Integer.parseInt(reader.nextLine());
+            double num1 = Integer.parseInt(reader.nextLine());
 
             System.out.print("Type another number: ");
-            int num2 = Integer.parseInt(reader.nextLine()); 
+            double num2 = Integer.parseInt(reader.nextLine());
 
 
             System.out.println("Division: " + num1 + "/" + num2 + "=" + num1/num2);
